+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and adheres to [Semantic Versioning](https://semver.org/).
 * Select larges tiles (do this logic). Talk with lab first.
 * Clean up file and do some testing
 
+#### Before production
+* Remove all logging and calculations done for logging i.e. pct of tiles indexed over possible tiles to index. Slows down the program.
+